+++ constants/main.go
@@ -3,23 +3,21 @@ package constants
 import "time"
 
 const (
-	ArpPacketPayloadSize     = 42
-	BuffersBurstLimit        = 4
+	EthIpv4ArpPacketSize     = 42
+	LimiterBuffersBurstLimit = 4
 	DefaultTimeout           = time.Second * 2
 	ErrorChanBufferSize      = 4
 	GatewayMacRequestTimeout = time.Second * 1
 	IOVecPacketsChunkSize    = 64
-	IcmpV4PacketPayloadSize  = 42
+	EthIpv4IcmpV4PacketSize  = 42
 	PcapCaptureTimeout       = time.Millisecond * 30
-	TcpV4PacketPayloadSize   = 54
-	UdpV4PacketPayloadSize   = 42
 	UpHostsChanSize          = 1024
 )
 
-const ArpAndEthernetHeadersSize = 22
+const EthernetAndArpHeadersSize = 22
 
-// ArpAndEthernetHeadersPart This array covers bytes [0..22) of the original skeleton (Ethernet + first part of ARP header).
-var ArpAndEthernetHeadersPart = [ArpAndEthernetHeadersSize]byte{
+// EthernetAndArpHeadersPart This array covers bytes [0..22) of the original skeleton (Ethernet + first part of ARP header).
+var EthernetAndArpHeadersPart = [EthernetAndArpHeadersSize]byte{
 	// Ethernet header (14 bytes)
 	0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, // [0:6]   Destination MAC (was [0:6])
 	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // [6:12]  Source MAC (was [6:12])
@@ -33,10 +31,10 @@ var ArpAndEthernetHeadersPart = [ArpAndEthernetHeadersSize]byte{
 	0x00, 0x01, // [20:22] Operation: 1 (ARP Request) (was [20:22])
 }
 
-const ArpPacketPayloadBodySize = 20
+const ArpPacketPayloadSize = 20
 
 // ArpPacketPayloadPart This array covers bytes [22..42) of the original skeleton (the ARP body).
-var ArpPacketPayloadPart = [ArpPacketPayloadBodySize]byte{
+var ArpPacketPayloadPart = [ArpPacketPayloadSize]byte{
 	// ARP body
 	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // [0:6]   Sender HW address (was [22:28])
 	0x00, 0x00, 0x00, 0x00, // [6:10]  Sender IP (was [28:32])
@@ -44,20 +42,14 @@ var ArpPacketPayloadPart = [ArpPacketPayloadBodySize]byte{
 	0x00, 0x00, 0x00, 0x00, // [16:20] Target IP (was [38:42])
 }
 
-const ArpPacketPaddingSize = 18
+const PacketEthernetV4PartSize = 14
 
-// ArpPacketPaddingPart This array covers bytes [42..60) of the original skeleton (the padding).
-var ArpPacketPaddingPart = [ArpPacketPaddingSize]byte{
-	// Padding (to reach the minimum Ethernet frame length of 60 bytes)
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
-	0x00, 0x00, 0x00, 0x00, 0x00, 0x00, // [0:18] (was [42:60])
-}
+const EighteenBytesPaddingPartSize = 18
 
-const ICMPPacketEthernetPartSize = 14
+var EighteenBytesPaddingPart = [EighteenBytesPaddingPartSize]byte{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0}
 
-// ICMPPacketEthernetPart is the Ethernet header (14 bytes).
-var ICMPPacketEthernetPart = [ICMPPacketEthernetPartSize]byte{
+// EthernetV4Part is the Ethernet header (14 bytes).
+var EthernetV4Part = [PacketEthernetV4PartSize]byte{
 	// [0:6]   Destination MAC - should be specified
 	0, 0, 0, 0, 0, 0,
 	// [6:12]  Source MAC - should be specified
@@ -66,12 +58,12 @@ var ICMPPacketEthernetPart = [ICMPPacketEthernetPartSize]byte{
 	8, 0,
 }
 
-const ICMPPacketIPPartSize = 20
+const IPv4PartSize = 20
 
-// ICMPPacketIPPart is the IPv4 header (20 bytes for a minimal header).
-var ICMPPacketIPPart = [ICMPPacketIPPartSize]byte{
-	// [0:4]   IPv4 Version, IHL, Type of Service, Total Length
-	69, 0, 0, 32,
+// IPv4TCPPart is the IPv4 header (20 bytes for a minimal header).
+var IPv4TCPPart = [IPv4PartSize]byte{
+	// [0:4]   IPv4 TCP Version, IHL, Type of Service, Total Length(should be specified)
+	69, 0, 0, 28,
 	// [4:6]   Identification
 	0, 0,
 	// [6:8]   Flags, Fragment Offset
@@ -84,10 +76,25 @@ var ICMPPacketIPPart = [ICMPPacketIPPartSize]byte{
 	255, 0, 0, 0,
 }
 
-const ICMPPacketICMPPartAndPaddingSize = 26
+var IPv4UDPPart = [IPv4PartSize]byte{
+	// [0:4]   IPv4UDP Version, IHL, Type of Service, Total Length
+	17, 0, 0, 16,
+	// [4:6]   Identification
+	0, 0,
+	// [6:8]   Flags, Fragment Offset
+	64, 0,
+	// [8:12]  TTL, Protocol (ICMP), Header Checksum
+	64, 1, 0, 0,
+	// [12:16] Source IP - should be specified
+	0, 0, 0, 0,
+	// [16:20] Destination IP - should be specified
+	255, 0, 0, 0,
+}
 
-// ICMPPacketICMPPartAndPadding is the ICMP header plus payload and padding (26 bytes).
-var ICMPPacketICMPPartAndPadding = [ICMPPacketICMPPartAndPaddingSize]byte{
+const IcmpV4PartSize = 26
+
+// IcmpV4Part is the ICMP header plus payload and padding (26 bytes).
+var IcmpV4Part = [IcmpV4PartSize]byte{
 	// [0:2]   ICMP Type (8 for Echo Request), Code (0)
 	8, 0,
 	// [2:4]   ICMP Checksum
@@ -96,8 +103,6 @@ var ICMPPacketICMPPartAndPadding = [ICMPPacketICMPPartAndPaddingSize]byte{
 	18, 52, 0, 1,
 	// [8:12]  Payload ("PING")
 	80, 73, 78, 71,
-	// [12:26] Padding to reach minimum Ethernet frame size
-	0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
 }
 
 // CommonPorts is a list of common service ports.
